Set TR doses for control to 0 in MIDAS
CNO does not tolerate multiple lines in MIDAS for dose at t=0.
This crashes CNO.
Former files which do not crash CNO were populated with 1s in the dose fields
for the control.
Although this runs, it does not make sense, unless the control was treated
with all drugs at t=0.
Thus, the control TR doses are set to 0.
Additionally, duplicate rows are dropped from the dataframe before saving
the MIDAS file.